Add a simple bar-graph animation as the music plays
Each bar is two tiles wide, and every two volume level is represented by
half a tile of height.

The top two volume levels are indicated with red, the two below that
amber, and the remaining levels green. The top segment of the bar graph
is given a highlight.

When updating the bar graph, code is in place to only write to the tiles
that are changing.