Populate pattern and colour tables for all three 'slices'
Previously, bit-masks were set up so that the whole screen can share a
single 256-entry pattern generator and colour table.

While this works on the SC-3000 with a real TMS99xx VDP, it does not work
on the Master System's backwards-compatibility mode-2.

To get something that works on both, just use three copies of the tables.
One each for the three screen slices.