Use Z80 interrupt-mode 2 instead of polling the VDP
Previously, polling of the VDP status register was used because
interrupt-mode 1 used a fixed address that we could not write to.

It turns out that polling seems to miss some interrupts, causing
slow-down is music playback. This may be the clear-on-read behaviour
of the status register clearing the interrupt before reporting it.

Instead, we can switch to using interrupt-mode 2. This allows us
to set the location of an interrupt vector table using the I register.
We do not have control the byte that indexes that table, so need to
populate 257 bytes with the same value, so that any index will give
the same vector.

The table is placed at 0x9600, and the repeated byte is 0x98, giving
an interrupt vector of 0x9898.

This fixes the slow-down issue that was being seen. With interrupts
now working, this also means we can switch to the normal version of
SGlib. The customized SGlib is no-longer needed as we don't need
to poll the VDP status register any more.